修正

+++ app/controllers/items_controller.rb
@@ -38,9 +38,11 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+   if @item.user_id == current_user.id
     if @item.destroy
       redirect_to root_path
     end
+   end
   end
 
   private
